+++ docs/src/dev-docs/changelog.rst
@@ -29,6 +29,8 @@ Fixed
 
 - ``NativePET`` can now evaluate on single-atom structures without crashing
 
+- The metatrain dataloader doesn't load all batches ahead of each epoch anymore
+
 Added
 #####
 
@@ -40,6 +42,9 @@ Changed
 
 - Training and exporting models without extensions will no longer lead to the creation
   of an empty directory for the extensions
+
+- The SOAP-BPNN model now uses ``torch-spex`` instead of ``featomic`` as its SOAP
+  backend
 
 Version 2025.5 - 2025-04-13
 ---------------------------